+++ src/validator/checker/mod.ts
@@ -1,0 +1,6 @@
+export * from "./array.ts";
+export * from "./optional.ts";
+export * from "./record.ts";
+export * from "./checker.ts";
+export * from "./string.ts";
+export * from "./number.ts";