+++ src/validator/type.ts
@@ -41,17 +41,15 @@ export type BasicType =
   | "object"
   | "function"
   | "null";
-/**
- * 元组项检测
- * @public
- */
-export type ExpectTypeTuple = ExpectType[];
+
+/**  @public */
+export type ExpectUnionType = ExpectType[];
 
 /**
  * 对象属性检测
  * @public
  */
-export type ExpectTypeObject = {
+export type ExpectObjectType = {
   [key: string | number]: ExpectType;
   [key: symbol]: any;
 };
@@ -67,8 +65,8 @@ export type ExpectType<T = unknown> =
   | TypeCheckFn<T>
   | TypeChecker<T>
   | BasicType
-  | ExpectTypeObject
-  | ExpectTypeTuple;
+  | ExpectObjectType
+  | ExpectUnionType;
 
 /** @public */
 export interface TypeCheckOptions {
@@ -91,12 +89,14 @@ export interface TypeCheckOptions {
 }
 type TypeCheckError = {
   error: TypeErrorDesc;
+  value?: undefined;
   replace?: undefined;
 };
 type TypeCheckReplace<T> = {
+  error?: undefined;
   /** 要替换的值 */
   value: T;
-  replace: true;
+  replace: boolean;
 };
 /** @public */
 export type TypeCheckFnCheckResult<T = unknown> =
@@ -121,7 +121,7 @@ type InferBaseMap = {
  * @public
  */
 export type InferExpect<T> = T extends string ? InferBaseMap[T]
-  : T extends any[] ? InferTuple<T>
+  : T extends any[] ? InferExpectUnion<T>
   : T extends TypeCheckFn<infer E> ? E
   : T extends TypeChecker<infer E> ? E
   : T extends object ? {
@@ -129,7 +129,8 @@ export type InferExpect<T> = T extends string ? InferBaseMap[T]
     }
   : unknown;
 
-type InferTuple<T extends any[]> = T extends [infer P, ...infer Q] ? [InferExpect<P>, ...InferTuple<Q>]
-  : T;
+/** @public */
+export type InferExpectUnion<T extends any[]> = T extends [infer P, ...infer Q] ? InferExpect<P> | InferExpectUnion<Q>
+  : never;
 
 type Fn = (...args: any[]) => any;