+++ test/validator/check.test.ts
@@ -1,16 +1,16 @@
-import { describe, expect, it } from "vitest";
+import { describe, expect, it, test } from "vitest";
 import { checkType } from "evlib/validator";
 import "./assests/type_check.assert.ts";
 describe("基础类型检测", function () {
   it("null", function () {
-    expect(checkType(null, "null")).toCheckPass();
-    expect(checkType(null, {})).toCheckFail();
+    expect(checkType(null, "null")).checkPass();
+    expect(checkType(null, {})).checkFail();
   });
   it("object", function () {
-    expect(checkType({}, "object")).toCheckPass();
+    expect(checkType({}, "object")).checkPass();
   });
   it("symbol", function () {
-    expect(checkType(Symbol(), "symbol")).toCheckPass();
+    expect(checkType(Symbol(), "symbol")).checkPass();
   });
 });
 describe("检测对象", function () {
@@ -18,7 +18,7 @@ describe("检测对象", function () {
     let obj = { s: 3, i: "s", q: undefined };
     expect(
       checkType(obj, { s: "number", i: "string", q: "undefined" }),
-    ).toCheckPass();
+    ).checkPass();
     expect(obj).toEqual({ s: 3, i: "s", q: undefined });
   });
   it("移除多余", function () {
@@ -28,7 +28,7 @@ describe("检测对象", function () {
       { s: "number", i: "string" },
       { policy: "delete" },
     );
-    expect(checkRes).toCheckPass();
+    expect(checkRes).checkPass();
     expect(obj).toEqual({ s: 3, i: "s" });
   });
   it("仅匹配", function () {
@@ -38,13 +38,13 @@ describe("检测对象", function () {
       { s: "number", i: "string" },
       { policy: "delete" },
     );
-    expect(checkRes).toCheckPass();
+    expect(checkRes).checkPass();
     expect(obj).toEqual({ s: 3, i: "s", q: undefined });
   });
   it("多余字段检测", function () {
     let obj = { s: 3, i: "s", q: undefined };
     let checkRes = checkType(obj, { s: "number", i: "string" });
-    expect(checkRes).toCheckFail({ q: "预期: 不存在, 实际: 存在" });
+    expect(checkRes).checkFail({ q: "预期: 不存在, 实际: 存在" });
     expect(obj).toEqual({ s: 3, i: "s", q: undefined });
   });
   it("检测所有字段", function () {
@@ -70,8 +70,8 @@ describe("检测对象", function () {
       q: "number",
       y: "number",
     });
-    expect(checkRes).toCheckFailWithField(["q"]);
-    expect(checkRes2).toCheckFailWithField(["q"]);
+    expect(checkRes).checkFailWithField(["q"]);
+    expect(checkRes2).checkFailWithField(["q"]);
     expect(obj).toEqual({ s: 3, i: "s", q: undefined });
   });
 
@@ -84,62 +84,33 @@ describe("检测对象", function () {
         { checkAll: true },
       ),
       "预期类型不一致",
-    ).toCheckFailWithField(["s", "y"]);
+    ).checkFailWithField(["s", "y"]);
     expect(
       checkType(obj, { s: "string", y: (a: any) => undefined }),
       "预期类型不一致",
-    ).toCheckFailWithField(["s"]);
+    ).checkFailWithField(["s"]);
   });
   it("预期不存在", function () {
     let res = checkType({ a: 8 }, { a: "number", b: "number" });
-    expect(res).toCheckFailWithField(["b"]);
+    expect(res).checkFailWithField(["b"]);
   });
   it("判断null类型", function () {
     let res = checkType({ a: null }, { a: "null" });
-    expect(res).toCheckPass();
+    expect(res).checkPass();
   });
   it("传入错误预期类型", function () {
     let res = checkType({ a: 3 }, { a: "D" } as any);
-    expect(res).toCheckFailWithField(["a"]);
+    expect(res).checkFailWithField(["a"]);
   });
 });
-describe("元组检测", function () {
-  it("全匹配", function () {
-    expect(checkType([1, "d"], ["number", "string"])).toCheckPass();
 
-    expect(checkType([1, "d"], ["number", "number"])).toCheckFailWithField([
-      "1",
-    ]);
-  });
-  it("长度检测", function () {
-    let val = [1, "d", null];
-    expect(checkType(val, ["number", "string"])).toCheckFail({
-      length: "预期长度: 2, 实际: 3",
-    });
-    expect(val).toEqual([1, "d", null]);
-  });
-  it("仅匹配预期提供字段", function () {
-    let val = [1, "d", null];
-    expect(
-      checkType(val, ["number", "string"], { policy: "pass" }),
-    ).toCheckPass();
-    expect(val).toEqual([1, "d", null]);
-  });
-  it("移除多余", function () {
-    let val = [1, "d", null];
-    expect(
-      checkType(val, ["number", "string"], { policy: "delete" }),
-    ).toCheckPass();
-    expect(val).toEqual([1, "d"]);
-  });
-});
 describe("嵌套", function () {
   it("仅检测", function () {
     let res = checkType(
       { s: 3, i: { q: "s", c: undefined } },
       { s: "number", i: { q: "string", c: "undefined" } },
     );
-    expect(res).toCheckPass();
+    expect(res).checkPass();
   });
   it("删除多余", function () {
     let obj = { s: 3, i: { q: "s", y: null, c: undefined }, b: 6 };
@@ -148,7 +119,19 @@ describe("嵌套", function () {
       { s: "number", i: { q: "string", c: "undefined" } },
       { policy: "delete" },
     );
-    expect(res).toCheckPass();
+    expect(res).checkPass();
     expect(obj).toEqual({ s: 3, i: { q: "s", c: undefined } });
   });
+});
+
+test("union", function () {
+  expect(checkType(null, ["string", "null"])).checkPass();
+  expect(checkType(undefined, ["string", "null"])).checkFail();
+  expect(checkType(
+    { s: 3, i: null },
+    { s: ["number", "string"], i: ["string", "null"] },
+  )).checkPass();
+  expect(
+    checkType({ s: 3 }, { s: ["bigint", "string"] }),
+  ).checkFailWithField(["s"]);
 });