+++ test/validator/checker/optional.test.ts
@@ -1,12 +1,12 @@
 import { describe, expect, test } from "vitest";
 import { checkType, optional, verifyType } from "evlib/validator";
-import "./assests/type_check.assert.ts";
+import "../assests/type_check.assert.ts";
 
 describe("自定义可选", function () {
   test("不存在的可选", function () {
     expect(
       checkType({ s: 3 }, { s: "number", q: optional("string") }),
-    ).toCheckPass();
+    ).checkPass();
   });
   test("错误的可选", function () {
     expect(
@@ -16,7 +16,7 @@ describe("自定义可选", function () {
           q: optional({ c: "string" }),
         },
       ),
-    ).toCheckFailWithField(["q"]);
+    ).checkFailWithField(["q"]);
   });
   test("正确的可选", function () {
     expect(
@@ -27,7 +27,7 @@ describe("自定义可选", function () {
           q: optional("string"),
         },
       ),
-    ).toCheckPass();
+    ).checkPass();
   });
   test("默认值", function () {
     expect(verifyType({}, { q: optional("string", undefined, 7) }).q).toEqual(7);
@@ -54,7 +54,7 @@ test("删除值为undefined且预期为可选类型的字段", function () {
       { s: "number", q: optional("string") },
       { policy: "delete" },
     ),
-  ).toCheckPass();
+  ).checkPass();
   expect(object, "q应该被删除").not.has.key("q");
 });
 test("快捷可选", function () {
@@ -63,17 +63,17 @@ test("快捷可选", function () {
       { s: 3, i: "s" },
       { s: "number", i: "string", q: optional.string },
     ),
-  ).toCheckPass();
+  ).checkPass();
   expect(
     checkType(
       { s: 3, i: "s", q: 8 },
       { s: "number", i: "string", q: optional.string },
     ),
-  ).toCheckFailWithField(["q"]);
+  ).checkFailWithField(["q"]);
   expect(
     checkType(
       { s: 3, i: "s", q: "sd" },
       { s: "number", i: "string", q: optional.string },
     ),
-  ).toCheckPass();
+  ).checkPass();
 });